requirements.txt
attrs==23.1.0
boxsdk==3.9.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
exceptiongroup==1.1.1
faker==19.12.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.29.0
requests-toolbelt==0.10.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
